tests/wpt/tests/web-share/canShare-insecure.http.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>WebShare Test: canShare from non-secure context</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
        test(() => {
          assert_false('canShare' in navigator, 'navigator has attribute \'canShare\'.');
        }, 'navigator.canShare must be undefined in non-secure context');
    </script>
  </body>
</html>
